Reorganizes internals
