+++ ffnn/include/ffnn/layer/impl/fully_connected.hpp
@@ -10,7 +10,7 @@
 #include <ffnn/assert.h>
 #include <ffnn/logging.h>
 #include <ffnn/optimizer/none.h>
-#include <ffnn/io/signature.h>
+#include <ffnn/internal/signature.h>
 
 namespace ffnn
 {
